+++ c/test.c
@@ -26,16 +26,16 @@
 #include "src/p0836.c"
 
 typedef struct {
-	unsigned int id;
-	unsigned long long answer;
-	unsigned long long (*func)();
+	uint32_t id;
+	uint64_t answer;
+	uint64_t (*func)();
 } Answer;
 
 const Answer answers[] = {
 	{1,		233168,			p0001},
 	{2,		4613732, 		p0002},
 	{3,		6857,			p0003},
-	{4,		906609,			(unsigned long long (*)()) p0004},
+	{4,		906609,			(uint64_t (*)()) p0004},
 	{5,		232792560,		p0005},
 	{6,		25164150,		p0006},
 	{7,		104743,			p0007},
@@ -53,7 +53,7 @@ const Answer answers[] = {
 	{25,	4782,			p0025},
 	{30,	443839,			p0030},
 	{34,	40730,			p0034},
-	{76,	190569291,		(unsigned long long (*)()) p0076},
+	{76,	190569291,		(uint64_t (*)()) p0076},
 };
 const size_t ANSWERS_LEN = sizeof(answers) / sizeof(answers[0]);
 static size_t current_index = 0;
@@ -64,7 +64,7 @@ void tearDown(void) {}
 
 void test_euler_answer() {
 	Answer key = answers[current_index];
-	unsigned long long result = key.func();
+	uint64_t result = key.func();
 	char *msg = (char*)malloc(256 * sizeof(char));
 	snprintf(msg, 256, "Euler problem %u should have an answer of %llu, but we actually got %llu", key.id, key.answer, result);
 	TEST_ASSERT_EQUAL_INT64_MESSAGE(key.answer, result, msg);
