+++ cplusplus/src/p0000_template.cpp
@@ -11,13 +11,13 @@ This paragraph should be replaced by the problem description, excluding images.
 #define EULER_P0000
 #include <stdio.h>
 
-unsigned long long p0000() {
+uint64_t p0000() {
     return 0;
 }
 
 #ifndef UNITY_END
 int main(int argc, char const *argv[]) {
-    unsigned long long answer = p0000();
+    uint64_t answer = p0000();
     printf("%llu\n", answer);
     return 0;
 }
